--- lab-car-service/java/src/dining/Dineable.java
@@ -1,5 +1,0 @@
-package dining;
-
-public interface Dineable {
-    void serveDinner(String carId);
-}
