--- lab-car-service/java/src/dining/RobotDinner.java
@@ -1,15 +1,0 @@
-package dining;
-
-public class RobotDinner implements Dineable{
-    private static int robotsServedCount = 0;
-
-    @Override
-    public void serveDinner(String carID){
-        System.out.println("Serving dinner to robots in car " + carID);
-        robotsServedCount++;
-    }
-
-    public static int getRobotsServed(){
-        return robotsServedCount;
-    }
-}
