--- lab-car-service/java/src/queue/Queue.java
@@ -1,10 +1,0 @@
-package queue;
-
-import model.Car;
-
-public interface Queue<T> {
-    void enqueue(T item);
-    T dequeue();
-    boolean isEmpty();
-    int size(); // Add the size method
-}
