--- lab-car-service/java/src/refueling/ElectricStation.java
@@ -1,15 +1,0 @@
-package refueling;
-
-public class ElectricStation implements Refuelable{
-    private static int electricCarsFueled = 0;
-
-    @Override
-    public void refuel(String carID){
-        System.out.println("Refueling electric car " + carID);
-        electricCarsFueled++;
-    }
-
-    public static int getElectricCarsFueled(){
-        return electricCarsFueled;
-    }
-}
