--- lab-car-service/java/src/refueling/Refuelable.java
@@ -1,4 +1,0 @@
-package refueling;
-
-public interface Refuelable {void refuel(String carId);
-}
