+++ lab-car-service/src/java/Cars/src/dining/Dineable.java
@@ -1,0 +1,5 @@
+package Cars.src.dining;
+
+public interface Dineable {
+    void serveDinner(String carId);
+}
