+++ lab-car-service/src/java/Cars/src/refueling/GasStation.java
@@ -1,0 +1,15 @@
+package Cars.src.refueling;
+
+public class GasStation implements Refuelable{
+    private static int gasCarsRefueled = 0;
+
+    @Override
+    public void refuel(String carID){
+        System.out.println("Gas cars refueled: " + carID);
+        gasCarsRefueled++;
+    }
+
+    public static int getGasCarsRefueled(){
+        return gasCarsRefueled;
+    }
+}
