+++ lab-car-service/src/java/Cars/src/refueling/Refuelable.java
@@ -1,0 +1,4 @@
+package Cars.src.refueling;
+
+public interface Refuelable {void refuel(String carId);
+}
