--- lab-more-coffee/task4/src/task4/main/Main.java
@@ -1,11 +1,0 @@
-package task4.main;
-
-import task4.barista.Barista; 
-
-public class Main {
-    public static void main(String[] args) {
-        Barista barista = new Barista();
-        barista.start(); // all interaction happens here
-    }
-}
-
